backend/app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
boto3==1.26.133
pydantic==1.10.7
mangum==0.17.0
python-dotenv==1.0.0

